requirements.txt
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
Jinja2==3.1.4
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
segfast==1.0.2
segyio==1.9.12
six==1.16.0
sympy==1.13.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
